backend/pyproject.toml
[project]
name = "personal-news-backend"
version = "0.1.0"
description = "Personal news dashboard backend"
authors = [
    {name = "hendrik-dev"}
]
requires-python = "^3.10"
dependencies = [
    "fastapi (==0.115.6)",
    "uvicorn[standard] (==0.34.0)",
    "feedparser (==6.0.11)",
    "newspaper4k (==0.9.3.1)",
    "httpx (==0.28.1)",
    "apscheduler (==3.10.4)",
    "pydantic (==2.10.6)",
    "python-dotenv (==1.0.1)",
    "lxml (==5.3.0)",
    "lxml-html-clean (==0.4.1)",
    "pillow (==11.1.0)",
    "aiofiles (==24.1.0)",
    "nltk (==3.9.1)"
]


[tool.poetry.scripts]
start = "uvicorn main:app --host 0.0.0.0 --port 8000 --reload"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
